desloppify: split shared-file convention outliers + skip ProcessRunner
- Split IPullRequestService.cs into 3 files (IPullRequestService.cs,
  IGitHubPullRequestService.cs, IGiteaPullRequestService.cs) so the
  two marker sub-interfaces no longer share a file with the canonical
  contract — matches the one-type-per-file convention everywhere else
  in Core/Services/.
- Split Ui/Converters/ToastConverters.cs into three sibling files
  (SeverityBrushConverter.cs, SeverityIconConverter.cs,
  ProgressToWidthConverter.cs), one IValueConverter per file. Matches
  the existing NullToVisibilityConverter / TabStatusToBrushConverter
  / ZeroToHidden / ZeroToVisible / InverseBoolToVisibility layout in
  the same folder.
- ProcessRunner partial-narrowing review: documented the intentional
  partial catch via wontfix attestation. Win32Exception/FileNotFound
  is the only case worth a typed user-visible Result; other exceptions
  bubble to the AppDomain unhandled handler so crash.log captures them.
