+++ src/CodeScope.Core/Services/IGiteaPullRequestService.cs
@@ -1,0 +1,7 @@
+namespace NoScope.CodeScope.Core.Services;
+
+/// <summary>
+/// Marker interface for the Gitea-backed provider (<c>tea</c> CLI). Role sibling of
+/// <see cref="IGitHubPullRequestService"/>.
+/// </summary>
+public interface IGiteaPullRequestService : IPullRequestService { }
